backend/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
email-validator==2.1.1

# Database
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# AI / RAG (Phase 5 — installed now, wired later)
langchain==0.2.3
langchain-community==0.2.3
langchain-openai==0.1.8
chromadb==0.5.0
openai==1.30.1
tiktoken==0.7.0

# Utilities
httpx==0.27.0
python-multipart==0.0.9
loguru==0.7.2
